requirements.txt
streamlit
requests
moviepy.editor import *
pytube import YouTube
tempfile
os
moviepy.editor import AudioFileClip
shutil
yt_dlp as youtube_dl